Develop task01

+++ bookstore-api/services/book.service.js
@@ -1,4 +1,6 @@
 import BookRepository from "../repositories/book.repository.js";
+import SaleRepository from "../repositories/sale.repository.js";
+import BookInfoRepository from "../repositories/bookInfo.repository.js";
 
 async function createBook(book) {
   await BookRepository.createBook(book);
@@ -9,23 +11,50 @@ async function updateBook(book) {
 }
 
 async function deleteBook(id) {
-  // const animal = await AnimalRepository.getAnimalByBookId(id);
-  // if (animal) {
-  //   throw new Error(
-  //     "Forbidden deletion! There are animals registered for this book!"
-  //   );
-  // } else {
-  //   return await BookRepository.deleteBook(id);
-  // }
-  return await BookRepository.deleteBook(id);
+  const sales = await SaleRepository.getSaleByBookId(id);
+  if (sales.length !== 0) {
+    throw new Error(
+      "Forbidden deletion! There are sales registered for this book!"
+    );
+  } else {
+    return await BookRepository.deleteBook(id);
+  }
 }
 
-async function getBooks() {
+async function getBooks(authorId) {
+  if (authorId) {
+    return await BookRepository.getBookByAuthorId(authorId);
+  }
   return await BookRepository.getBooks();
 }
 
 async function getBook(id) {
-  return await BookRepository.getBook(id);
+  const book = await BookRepository.getBook(id);
+  book.info = await BookInfoRepository.getBookInfo(parseInt(id));
+  const bookComplete = book.dataValues;
+  bookComplete.info = book.info;
+  // console.log("bookComplete: ", bookComplete);
+  return bookComplete;
+}
+
+async function createBookInfo(bookInfo) {
+  await BookInfoRepository.createBookInfo(bookInfo);
+}
+
+async function updateBookInfo(bookInfo) {
+  await BookInfoRepository.updateBookInfo(bookInfo);
+}
+
+async function deleteBookInfo(bookId) {
+  await BookInfoRepository.deleteBookInfo(bookId);
+}
+
+async function createReview(review, bookId) {
+  await BookInfoRepository.createReview(review, bookId);
+}
+
+async function deleteReview(bookId, index) {
+  await BookInfoRepository.deleteReview(parseInt(bookId), index);
 }
 
 export default {
@@ -34,4 +63,9 @@ export default {
   deleteBook,
   getBooks,
   getBook,
+  createBookInfo,
+  updateBookInfo,
+  deleteBookInfo,
+  createReview,
+  deleteReview,
 };
